Shipping build targets. Spent most of the day trying to track down errors from App Verifier that just turned out to be normal heap exhaustion from PageHeap. Builds are squeeky clean with 64 bit App Verifier builds.
App Verifier did lead me to catch an issue with my RAD_STRING_DISABLE_STACK_STRINGS_FOR_REFTAG_DATA option in the string class. I hadn't wrapped all the appropriately places with that and it was causing string conversions to fail in a non-deterministic and ultimately crashy sort of way.

Signed-off-by: joer <joer@sunsidegames.com>